pyproject.toml
[tool.poetry]
name = "product-recommendation-system"
version = "0.1.0"
description = "This tool generates recommendations using state of art AI algorithms"
authors = ["Usman Mushtaq <usmanmushtaq1990@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.3.3"
azureml-core = "^1.34.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
set-pythonpath = "sh -c 'export PYTHONPATH=${PYTHONPATH}:$(pwd)/src'"

[tool.pytest.ini_options]
python_paths = ["src"]
